requirements.txt
build~=1.0.3
coverage~=7.3.1
flake8~=6.1.0
mypy~=1.5.1
python-periphery~=2.4.1
Sphinx~=7.2.6
twine~=4.0.2
